Update results_tracking_60min.ipynb
I modified the code to deal with a vector of phasor targets (ex. for vmag) that was the same length as the actual vmag vector.
In order for this version of code to work correctly, the simulink simulation must have 3600 seconds (the entire hour), and cannot be run at a portion of the 60 minutes.